+++ setup.py
@@ -8,5 +8,5 @@ setup(
     install_requires=["pandas", "pysam"],
     author="Dillon O.R. Barker",
     author_email="dillon.barker@canada.ca",
-    entry_points={"console_scripts": ["mmmvi=mmmvi:main"]},
+    entry_points={"console_scripts": ["mmmvi=mmmvi.mmmvi:main"]},
 )
